Fixed some integer length issues.

+++ ext/oily_png/png_decoding.h
@@ -4,7 +4,7 @@
 #define BUILD_PIXEL(r, g, b, a) (((PIXEL) r << 24) + ((PIXEL) g << 16) + ((PIXEL) b << 8) + (PIXEL) a)
 #define UNFILTER_BYTE(byte, adjustment)  byte = (BYTE) (((byte) + (adjustment)) & 0x000000ff)
 
-typedef PIXEL(*pixel_decoder_func)(BYTE*, int, int, VALUE);
+typedef PIXEL(*pixel_decoder_func)(BYTE*, long, long, VALUE);
 
 
 /*
